+++ package.json
@@ -15,7 +15,7 @@
     "eslint": "eslint src",
     "eslint-fix": "npm run eslint -- --fix",
     "prepublishOnly": "npm run compile",
-    "test": "npm run eslint",
+    "test": "jest && npm run eslint",
     "test-only": "jest",
     "test-coverage": "jest --coverage"
   },
